mqtt-service-python/requirements.txt
awscrt==0.27.4
AWSIoTPythonSDK==1.5.4
awsiotsdk @ file:///Users/kakf/Projects/Zea.Test/aws-iot-device-sdk-python-v2
source==1.2.0
